Add the start of a C# API wrapper.
This also includes a couple of examples based on .NET core and WPF. It has only been built and tested using Visual Studo 2017, but I think it should work with 2015 as well. The WPF example is currently broken due to a yet unknown dependency issue, but I figured it was worth including anyway.