Add the X-Adi-Application-Name header.
This also moves the header code to EmsApiServiceConfiguration, and removes the duplicate HttpClient being instantiated in the authentication handler (it can use the main HttpClient instead by calling SendAsync). I was hoping this would make it so we only need to set the headers in one place, but since the auth handler creates the request from scratch, it still needs them applied, because HttpClient.DefaultRequestHeaders does not come into play.